feat(nnue): setup NNUE accumulator structures and TODOs
- Upgraded Accumulator arrays to int16_t for AVX2 SIMD optimizations
- Outlined Network load and evaluate structure with clear TODOs
- Integrated Accumulator into GameState for instant restoration on unmakeMove
- Placed explicit TODO hooks in BoardMove.cpp for piece feature updates

+++ include/eval/nnue/Network.hpp
@@ -10,13 +10,18 @@ namespace VEngine {
     
     class Network {
     public:
+        // TODO: Implement the dense layer forward pass inside Network::evaluate() 
+        // to collapse the 256-value accumulator into a single centipawn score.
         static int evaluate(const Board& board);
         static int lazyEvaluate(const Board& board);
+        
+        // TODO: Write the Network::load() function to actually read a .nnue binary file.
         static bool load(const std::string& path);
-        static const std::vector<float>& getBiases();
+        static const std::vector<int16_t>& getBiases();
 
         // NNUE specific
-        static void computeLayer(float* input, float* weights, float* biases, float* output, int inSize, int outSize, bool activate);
+        // TODO: Update layer computation to use int16_t/int8_t SIMD
+        static void computeLayer(int16_t* input, int16_t* weights, int32_t* biases, int32_t* output, int inSize, int outSize, bool activate);
     }; 
 }
